m4/host-os.m4: Fixed bug inserting string 'GNU/' to Linux, kFreeBSD, etc
Operating systems consist of far more than their core utilities. If its
argued that we should acknowledge the userspace, we should be consistent
and acknowledge all of the userspace. Change the string to
'BSD/Oracle/Microsoft/Redhat/ISC/Apache/MIT/GNU/Linux'. This would quickly
grow unwieldly and ridiculous, and thus should be avoided. Additionally,
installing a single GNU utility such as 'uname' on a FreeBSD machine, does
not make the system GNU/FreeBSD.

Sane alternatives to this bug fix may be:
* Or replace the 'o' flag with an attempt to determine the actual OS.
  (ie, Fedora, Ubuntu, Arch, etc.) from standard version files.
* Update uname(3) and 'struct utsname' to include OS version information.
* Just call it Linux.